CORDA-2672: Refactor Attachment Demo.
