+++ samples/attachment-demo/build.gradle
@@ -1,17 +1,60 @@
 apply plugin: 'kotlin'
 apply plugin: 'idea'
 apply plugin: 'net.corda.plugins.quasar-utils'
+apply plugin: 'net.corda.plugins.cordapp'
 apply plugin: 'net.corda.plugins.cordformation'
 
 description 'Corda attachment demo'
 
+cordapp {
+    info {
+        name "Corda Attachment Demo"
+        vendor "R3"
+        targetPlatformVersion corda_platform_version.toInteger()
+        minimumPlatformVersion 1
+    }
+}
+
+sourceSets {
+    integrationTest {
+        kotlin {
+            compileClasspath += main.output + test.output
+            runtimeClasspath += main.output + test.output
+            srcDir file('src/integration-test/kotlin')
+        }
+    }
+}
+
+configurations {
+    integrationTestCompile.extendsFrom testCompile
+    integrationTestRuntime.extendsFrom testRuntime
+}
+
 dependencies {
+    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    compile "net.sf.jopt-simple:jopt-simple:$jopt_simple_version"
+    compile "javax.servlet:javax.servlet-api:3.1.0"
+    compile "javax.ws.rs:javax.ws.rs-api:2.0.1"
+    cordaCompile project(':client:rpc')
+    runtimeOnly "org.slf4j:slf4j-simple:$slf4j_version"
+
     // Corda integration dependencies
     cordaRuntime project(path: ":node:capsule", configuration: 'runtimeArtifacts')
     cordaRuntime project(path: ":webserver:webcapsule", configuration: 'runtimeArtifacts')
 
     cordapp project(':samples:attachment-demo:contracts')
     cordapp project(':samples:attachment-demo:workflows')
+
+    testCompile project(':node-driver')
+    testCompile "junit:junit:$junit_version"
+    testCompile "org.assertj:assertj-core:$assertj_version"
+    
+    integrationTestCompile project(':webserver')
+}
+
+task integrationTest(type: Test, dependsOn: []) {
+    testClassesDirs = sourceSets.integrationTest.output.classesDirs
+    classpath = sourceSets.integrationTest.runtimeClasspath
 }
 
 def nodeTask = tasks.getByPath(':node:capsule:assemble')
@@ -69,16 +112,16 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar', nodeTask, 
     }
 }
 
-task runSender(type: JavaExec) {
+task runSender(type: JavaExec, dependsOn: jar) {
     classpath = sourceSets.main.runtimeClasspath
-    main = 'net.corda.attachmentdemo.workflows.AttachmentDemoKt'
+    main = 'net.corda.attachmentdemo.AttachmentDemoKt'
     args '--role'
     args 'SENDER'
 }
 
-task runRecipient(type: JavaExec) {
+task runRecipient(type: JavaExec, dependsOn: jar) {
     classpath = sourceSets.main.runtimeClasspath
-    main = 'net.corda.attachmentdemo.workflows.AttachmentDemoKt'
+    main = 'net.corda.attachmentdemo.AttachmentDemoKt'
     args '--role'
     args 'RECIPIENT'
 }
