+++ samples/attachment-demo/src/main/resources/bank-of-london-cp.jar
